Homogenized shops code with products code

+++ backend/app/resources/shops.py
@@ -1,12 +1,11 @@
 import itertools
-from datetime import date
 from flask_restful import Resource
 from flask import request
 from webargs import fields
 from webargs.flaskparser import use_args
 from marshmallow import validate, ValidationError
 from marshmallow.decorators import post_dump, pre_dump
-from app.models import Product, Shop,ShopTag, Price, db, ma
+from app.models import Shop,ShopTag, db, ma
 from sqlalchemy import asc, desc
 from app.resources.auth import requires_auth
 
@@ -38,8 +37,8 @@ shop_schema = ShopSchema()
 
 class ShopsResource(Resource):
     @use_args({
-        'start': fields.Int(missing=1, location='query'),
-        'count': fields.Int(missing=20, location='query'),
+        'start': fields.Int(missing=0, location='query',validate=validate.Range(min=0)),
+        'count': fields.Int(missing=20, location='query',validate=validate.Range(min=0)),
         'sort': fields.Str(missing='id|ASC', location='query',
                            many=True, validate=validate.OneOf(SORT_CHOICE)),
         'status': fields.Str(missing='ACTIVE',location='query',validate=validate.OneOf(STATUS_CHOICE)),
@@ -47,7 +46,7 @@ class ShopsResource(Resource):
     })
     def get(self, args):
         query = Shop.query
-        start = args['start']
+        start = args['start']+1
         count = args['count']
         status = args['status']
         sort = {
@@ -62,22 +61,21 @@ class ShopsResource(Resource):
         shops_page = query.paginate(start, count)
         shops = shop_schema.dump(shops_page.items,many=True).data
         return {
-            'start': start,
-            'count': count,
+            'start': shops_page.page-1,
+            'count': len(shops_page.items),
             'total': shops_page.total,
             'shops': shops
         }
-
+    @requires_auth
     @use_args({
-        'name': fields.String(required=True, location='json'),
-        'address': fields.String(required=True, location='json'),
-        'lng': fields.Float(required=True,location='json'),
-        'lat': fields.Float(required=True,location='json'),
-        'tags': fields.List(fields.String(),required=True,location='json'),
+        'name': fields.Str(required=True, location='form'),
+        'address': fields.Str(required=True, location='form'),
+        'lng': fields.Float(required=True,location='form'),
+        'lat': fields.Float(required=True,location='form'),
+        'tags': fields.List(fields.String(),required=True,location='form'),
         'format': fields.Str(missing='json',location='query', validate=validate.Equal('json'))
     })
     def post(self, args):
-        del args['format']
         new_shop = Shop(name=args['name'],address=args['address'],lng=args['lng'],lat=args['lat'], withdrawn=False)
         new_shop.tags = [ShopTag(name=tag, shop=new_shop) for tag in args['tags'] if tag.strip()]
         db.session.add(new_shop)
@@ -91,13 +89,14 @@ class ShopResource(Resource):
     def get(self, args, shop_id):
         shop = Shop.query.get_or_404(shop_id)
         return shop_schema.dump(shop).data
-
+    
+    @requires_auth
     @use_args({
-        'name': fields.String(required=True, location='json'),
-        'address': fields.String(required=True, location='json'),
-        'lng': fields.Float(required=True,location='json'),
-        'lat': fields.Float(required=True,location='json'),
-        'tags': fields.List(fields.Str(),required=True,location='json'),
+        'name': fields.Str(required=True, location='form'),
+        'address': fields.Str(required=True, location='form'),
+        'lng': fields.Float(required=True,location='form'),
+        'lat': fields.Float(required=True,location='form'),
+        'tags': fields.List(fields.Str(),required=True,location='form'),
         'format': fields.Str(location='query', validate=validate.Equal('json'))
     })
     def put(self, args, shop_id):
@@ -111,13 +110,14 @@ class ShopResource(Resource):
         shop.tags = [ShopTag(name=tag, shop=shop) for tag in args['tags'] if tag.strip()]
         db.session.commit()
         return shop_schema.dump(shop).data
-
+    
+    @requires_auth
     @use_args({
-        'name': fields.String(location='json'),
-        'address': fields.String(location='json'),
-        'lng': fields.Float(location='json'),
-        'lat': fields.Float(location='json'),
-        'tags': fields.List(fields.Str(),location='json'),
+        'name': fields.Str(location='form'),
+        'address': fields.Str(location='form'),
+        'lng': fields.Float(location='form'),
+        'lat': fields.Float(location='form'),
+        'tags': fields.List(fields.Str(),location='form'),
         'format': fields.Str(missing='json', location='query', validate=validate.Equal('json'))
     })
     def patch(self, args, shop_id):
@@ -139,11 +139,13 @@ class ShopResource(Resource):
     @use_args({
         'format': fields.Str(missing='json', location='query', validate=validate.Equal('json'))
     })
-    def delete(self, _args, shop_id,is_admin):
-        shop = Shop.query.get_or_404(shop_id)
+    def delete(self, _args, shop_id,is_admin, **_kwargs):
+        shop = Shop.query.get(shop_id)
+        if not(shop):
+            return {'message': "Shop not found on DB"},404 
         if is_admin:
             db.session.delete(shop)
         else:
             shop.withdrawn = True
         db.session.commit()
-        return {'message', 'OK'}
+        return {'message': 'OK'}
